fix some bugs

+++ src/components/image-upload/ImagesUpload.vue
@@ -34,20 +34,10 @@
                 @click="handlePreview(file)">
               <el-button type="primary" size="mini" icon="el-icon-zoom-in" circle></el-button>
             </span>
-            <span
-                class="el-upload-list__item-delete"
-                @click="handleDownload(file)">
-              <el-button type="primary" size="mini" icon="el-icon-download" circle></el-button>
-            </span>
-            <span
-                class="el-upload-list__item-delete"
-                @click="handleRemove(file)">
-              <el-button type="danger" size="mini" icon="el-icon-delete" circle></el-button>
-            </span>
           </span>
           <div
               style="width: 100%;background: #fff;border-top: 1px solid #c0ccda;position:absolute;margin-left: 5px;bottom: 0;text-align: left">
-            <el-checkbox v-model="list" @mouseenter="handleDownload" :label="file.id">
+            <el-checkbox v-model="list" :label="file.id">
               {{ file.name.slice(22) }}
             </el-checkbox>
           </div>
@@ -127,20 +117,32 @@ export default {
       //
     },
     async handleRemove(file) {
-      const res = await this.$http.delete('/picture/delete/' + file.id)
-      if (res.status) {
-        this.$message.success(res.message)
-        let index = this.fileList.findIndex(item => item.id === file.id)
-        this.fileList.splice(index, 1)
-        this.sourceImageObjects.splice(index, 1)
-        return
+      if (file.id) {
+        const res = await this.$http.delete('/picture/delete/' + file.id)
+        if (res.status) {
+          this.$message.success(res.message)
+          let index = this.fileList.findIndex(item => item.id === file.id)
+          this.fileList.splice(index, 1)
+          this.sourceImageObjects.splice(index, 1)
+          return
+        }
+        this.$message.error(res.message)
       }
-      this.$message.error(res.message)
+    },
+    handleRemoveBatch(listIds) {
+      listIds.forEach(id => {
+        let index = this.fileList.findIndex(item => item.id === id)
+        if (index > -1) {
+          this.fileList.splice(index, 1)
+          this.sourceImageObjects.splice(index, 1)
+        }
+      })
     },
     handlePreview(_file) {
       let index = this.sourceImageObjects.findIndex(k => k.src === _file.url) | 0
       viewerApi({
         options: {
+          zIndex: 20016,
           toolbar: true,
           url: 'data-source',
           initialViewIndex: index
@@ -152,8 +154,33 @@ export default {
 
     },
     submitUpload() {
-      this.$refs.upload.submit()
       let length = this.uploadFileList.length
+      if (length > 10) {
+        this.$notify({
+          title: '警告',
+          message: '单次上传图片数限制为10',
+          type: 'warning'
+        })
+        return
+      }
+      const noMatchImagesSuffix = ["image/jpeg", "image/png", "image/jpg"]
+      const noMatchImages = []
+      this.uploadFileList.forEach(item => {
+        if (!noMatchImagesSuffix.includes(item.type)) {
+          noMatchImages.push(item.name)
+        }
+      })
+      if (noMatchImages.length > 0) {
+        this.$notify({
+          dangerouslyUseHTMLString: true,
+          title: '仅支持jpeg、png、jpg格式的图片文件',
+          message: `${noMatchImages.map((item, index) => index + 1 + '：' + item + '<br>').toString().replaceAll(',', '')}`,
+          type: 'warning',
+          duration: 0
+        })
+        return
+      }
+      this.$refs.upload.submit()
       return this.$http.post(`/file/upload/${this.namespace}/${this.type}/${this.resourceId}`
           , this.formData).then(res => {
             if (res.status) {
@@ -178,10 +205,6 @@ export default {
       ).catch(res => {
         console.log(res)
       })
-    },
-
-    handleDownload(file) {
-      window.open("http://127.0.0.1:8848/api/file/download/" + file.id, '_self')
     },
   },
   watch: {
